+++ Cargo.toml
@@ -18,7 +18,7 @@ members = [
   "query-engine/connectors/mongodb-query-connector",
   "query-engine/core",
   "query-engine/query-engine",
-  "query-engine/query-engine-napi",
+  "query-engine/query-engine-node-api",
   "query-engine/connector-test-kit-rs/query-engine-tests",
   "query-engine/request-handlers",
   "prisma-fmt",
