--- query-engine/connector-test-kit-rs/query-tests-setup/src/runner/napi.rs
@@ -1,1 +1,0 @@
-pub struct NApiRunner {}
