+++ query-engine/connector-test-kit-rs/query-tests-setup/src/runner/node_api.rs
@@ -1,0 +1,1 @@
+pub struct NodeApiRunner {}
