+++ query-engine/connectors/sql-query-connector/src/database/mod.rs
@@ -27,7 +27,7 @@ pub trait FromSource {
     ///    vars and validating, which can fail with a schema parser error. We
     ///    want to handle this as early as possible and in a single place.
     ///
-    /// 2. The url may be modified with the config dir, in the case of NAPI.
+    /// 2. The url may be modified with the config dir, in the case of Node-API.
     async fn from_source(source: &Datasource, url: &str) -> connector_interface::Result<Self>
     where
         Self: Connector + Sized;
