+++ query-engine/query-engine-node-api/src/error.rs
