+++ query-engine/query-engine-node-api/src/lib.rs
@@ -3,7 +3,7 @@
 mod engine;
 mod error;
 mod logger;
-mod napi_generators;
+mod node_api;
 
 pub(crate) type Result<T> = std::result::Result<T, error::ApiError>;
 pub(crate) type Executor = Box<dyn query_core::QueryExecutor + Send + Sync>;
