+++ query-engine/query-engine-node-api/src/logger/registry.rs
