+++ query-engine/query-engine-node-api/src/logger/telemetry.rs
