+++ query-engine/query-engine-node-api/src/logger/visitor.rs
