+++ query-engine/query-engine-node-api/src/node_api/functions.rs
